+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `sarkit.cphd.Reader` support for reading a subset of vectors
+
 
 ## [1.0.1] - 2025-09-09